split one file into two
